Milestone project 2
This is a fun project from the course "Complete Python Bootcamp From Zero to Hero in Python" hosted by José Portilla on Udemy website. The goal was to create a simplified version of the "Blackjack" game. The rules for creation and coding are placed on the first cell of the notebook file.